+++ controls/src/real_time_daemon.cpp
@@ -302,12 +302,12 @@ void RealTimeDaemon::handleCommand(const std::string& jsonStr)
         } else if (cmd == "setSingleAngle") {
             int spin = root.get("spinDirection", 0).asInt();
             int angle = root.get("angle", 0).asInt();
-            mot.setSingleAngle(static_cast<uint8_t>(spin), static_cast<uint16_t>(angle));
+            mot.setSingleAngle(static_cast<uint8_t>(spin), static_cast<int32_t>(angle));
         } else if (cmd == "setSingleAngleWithSpeed") {
             int spin = root.get("spinDirection", 0).asInt();
             int angle = root.get("angle", 0).asInt();
             int maxSpeed = root.get("maxSpeed", 0).asInt();
-            mot.setSingleAngleWithSpeed(static_cast<uint8_t>(spin), static_cast<uint16_t>(angle), static_cast<uint16_t>(maxSpeed));
+            mot.setSingleAngleWithSpeed(static_cast<uint8_t>(spin), static_cast<int32_t>(angle), static_cast<uint16_t>(maxSpeed));
         } else if (cmd == "setIncrementAngle") {
             int value = root.get("incAngle", 0).asInt();
             mot.setIncrementAngle(static_cast<int32_t>(value));
